requirements.txt
django==5.0.3
djangorestframework==3.14.0
python-dotenv==1.0.0
python-dateutil==2.8.2
